### Changed - 会員リポジトリインスタンスにユーザープールIDを持たせるように変更

+++ src/service/task/deleteMember.ts
@@ -4,6 +4,7 @@ import * as factory from '../../factory';
 
 import { MongoRepository as ActionRepo } from '../../repo/action';
 import { CognitoRepository as PersonRepo } from '../../repo/person';
+import { MongoRepository as ProjectRepo } from '../../repo/project';
 import { MongoRepository as TaskRepo } from '../../repo/task';
 
 import * as CustomerService from '../customer';
@@ -13,9 +14,21 @@ import * as CustomerService from '../customer';
  */
 export function call(data: factory.task.IData<factory.taskName.DeleteMember>): IOperation<void> {
     return async (settings: IConnectionSettings) => {
+        const projectRepo = new ProjectRepo(settings.connection);
+        const projectId = (data.project !== undefined) ? data.project.id : <string>process.env.PROJECT_ID;
+        const project = await projectRepo.findById({ id: projectId });
+        if (project.settings === undefined
+            || project.settings.cognito === undefined) {
+            throw new factory.errors.ServiceUnavailable('Project settings undefined');
+        }
+
+        const personRepo = new PersonRepo({
+            userPoolId: project.settings.cognito.customerUserPool.id
+        });
+
         await CustomerService.deleteMember(data)({
             action: new ActionRepo(settings.connection),
-            person: new PersonRepo(),
+            person: personRepo,
             task: new TaskRepo(settings.connection)
         });
     };
